requirements.txt
openai>=1.30.0
anthropic>=0.25.0
python-docx>=1.1.0
pdfplumber>=0.10.0
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0